Move to src dir layout
This makes it easier to know what version of the code is being tested,
because it has to be installed into an environment for the tests to run.
Although not officially recommended by PyPA, the discussion is moving
towards that outcome. See pypa/packaging.python.org#320
