+++ .coveragerc
@@ -10,7 +10,11 @@ exclude_lines =
     if __name__ == .__main__.:
 ignore_errors = True
 omit =
-    thermostate/tests/*
     setup.py
-    thermostate/_version.py
+    src/thermostate/_version.py
     .eggs/*
+
+[paths]
+source =
+   src/thermostate
+   .tox/*/lib/python*/site-packages/thermostate
