+++ setup.py
@@ -57,7 +57,8 @@ setup(
         'Operating System :: POSIX',
         'Operating System :: POSIX :: Linux',
     ],
-    packages=find_packages(),
+    packages=["thermostate"],
+    package_dir={"": "src"},
     install_requires=install_requires,
     tests_require=tests_require,
     setup_requires=setup_requires,
