+++ src/thermostate/__init__.py
@@ -1,3 +1,3 @@
 from .thermostate import units, Q_, State  # noqa
 from .abbreviations import EnglishEngineering, SystemInternational  # noqa
-from ._version import __version__
+from ._version import __version__  # noqa
